updated members location
