+++ react-app/src/components/ChannelsPage/index.js
@@ -4,6 +4,8 @@ import { Link, useParams } from 'react-router-dom'
 import { getAllChannels } from '../../store/channel';
 import { useHistory, Redirect } from 'react-router';
 import * as serverActions from '../../store/server';
+import * as conversationActions from '../../store/conversation';
+
 
 
 
@@ -19,12 +21,22 @@ function ChannelsList() {
     let [users, setUsers] = useState(null)
     let [displayUsers, setDisplayUsers] = useState(false)
     const currentUser = session?.user.id
+    let [members, setMembers] = useState(null)
 
+    
+    async function fetchData() {
+        const res = await fetch(`/api/users/members/${serverId}`)
+        const responseData = await res.json()
+        console.log(responseData)
+        setMembers(responseData.users)
+    }
 
-    const joinServerButton = ({userId}) => {
+    const joinServerButton = async ({userId}) => {
         setUsers(null)
         setDisplayUsers(false)
-        return dispatch(serverActions.joinAServer({ userId, serverId }))
+        await dispatch(serverActions.joinAServer({ userId, serverId }))
+        fetchData()
+        return
     }
 
     const joinServerAdminButton = () => {
@@ -41,6 +53,7 @@ function ChannelsList() {
     useEffect(() => {
         if (serverId) {
             dispatch(getAllChannels(serverId))
+            fetchData()
         }
 
         //if server state is empty, return them to the servers page
@@ -54,7 +67,6 @@ function ChannelsList() {
             const response = await fetch(`/api/users/${searchValue}`);
             const responseData = await response.json();
             setUsers(responseData.users)
-            console.log(users)
         }
         if(!(/^\s+$/.test(searchValue)) && searchValue !== ''){
             fetchData();
@@ -82,67 +94,68 @@ function ChannelsList() {
         }
     }
 
-    const test = (
-        <p>hello</p>
-    )
-
-    const hi = () => {
-        if (users) {
-            const display = users.forEach(user => {
-                console.log(user)
-                return (
-                    <p>Hello</p>
-                )
-            })
-            
-            return display
-        }
+    const startConversation = async ({userId}) => {
+        let res = await dispatch(conversationActions.addNewConversation({from_user:currentUser, to_user:userId}))
+        history.push(`/servers/0/conversations/${res.id}/messages`)
     }
 
     return (
-        <div id='channels_container'>
-            <h1>Channels:</h1>
-            {serverSelected &&
-                channels.map(channel =>
-                    <div key={channel.id}>
-                        <h2><Link to={`/servers/${serverId}/channels/${channel.id}/messages`}>{channel.channel_name}</Link></h2>
-                        <Link to={`/servers/${serverId}/channels/${channel.id}/edit`}>Edit</Link>
+        <>
+            <div id='channels_container'>
+                <h1>Channels:</h1>
+                {serverSelected &&
+                    channels.map(channel =>
+                        <div key={channel.id}>
+                            <h2><Link to={`/servers/${serverId}/channels/${channel.id}/messages`}>{channel.channel_name}</Link></h2>
+                            <Link to={`/servers/${serverId}/channels/${channel.id}/edit`}>Edit</Link>
+                        </div>
+                    )}
+                {serverSelected && <Link to={`/servers/${serverId}/channels/new`}>Add A Channel</Link>}
+                {!serverSelected && <h3>Select A Server</h3>}
+
+
+                <div className="serverOptions">
+                    {serverSelected &&
+                        <>  
+                            { displayUsers && (
+                                <input type='text' onChange={e => setSearchValue(e.target.value)} value={searchValue}></input>
+                            )}
+                            <button onClick={() => {
+                                setDisplayUsers(!displayUsers)
+                                setSearchValue('')
+                            }}>Add Server Member</button>
+                            {hideButton() === true &&
+                                <button onClick={joinServerAdminButton}>Join as Admin!</button>
+                            }
+                            <button><Link to={`/servers/edit/${serverId}`} style={{color:'black'}}>Edit Server</Link></button>
+                            <button onClick={() => sendId(serverId)}>Delete Server</button>
+                        </>
+                    }
+                </div>
+                {}
+                {users && users.map((user) => 
+                    <div>
+                        <div key={user.id} className="server_info_block">
+                            <a onClick={() => joinServerButton({userId: user.id})} className='server_a'>
+                                <img className={`server_image ${user.id === parseInt(serverId) ? 'selected' : ''}`} src={user.profile_image} alt={user.username} /></a>
+                            <p>{`${user.username}`}</p>
+                        </div>
                     </div>
                 )}
-            {serverSelected && <Link to={`/servers/${serverId}/channels/new`}>Add A Channel</Link>}
-            {!serverSelected && <h3>Select A Server</h3>}
 
-
-            <div className="serverOptions">
-                {serverSelected &&
-                    <>  
-                        { displayUsers && (
-                            <input type='text' onChange={e => setSearchValue(e.target.value)} value={searchValue}></input>
-                        )}
-                        <button onClick={() => {
-                            setDisplayUsers(!displayUsers)
-                            setSearchValue('')
-                        }}>Add Server Member</button>
-                        {hideButton() === true &&
-                            <button onClick={joinServerAdminButton}>Join as Admin!</button>
-                        }
-                        <button><Link to={`/servers/edit/${serverId}`} style={{color:'black'}}>Edit Server</Link></button>
-                        <button onClick={() => sendId(serverId)}>Delete Server</button>
-                    </>
-                }
             </div>
-            {}
-            {users && users.map((user) => 
-                <div>
-                    <div key={user.id} className="server_info_block">
-                        <a onClick={() => joinServerButton({userId: user.id})} className='server_a'>
-                            <img className={`server_image ${user.id === parseInt(serverId) ? 'selected' : ''}`} src={user.profile_image} alt={user.username} /></a>
-                        <p>{`${user.username}`}</p>
+            <div id='members_container'>
+                {members && members.map((user) => 
+                    <div>
+                        <div key={user.id} className="server_info_block">
+                            <a onClick={() => startConversation({userId: user.id})} className='server_a'>
+                                <img className={`server_image`} src={user.profile_image} alt={user.username} /></a>
+                            <p>{`${user.username}`}</p>
+                        </div>
                     </div>
-                </div>
-            )}
-
-        </div>
+                )}
+            </div>
+        </>
     )
 }
